Attempt 1 at fix for duped ACK handling.

+++ includes/mytftp/types.hpp
@@ -48,7 +48,7 @@ namespace TftpServer::MyTftp {
     };
 
     struct AckPayload {
-        unsigned int block_n;
+        unsigned short block_n;
     };
 
     struct ErrorPayload {
